Update TwitterUsers.scala
Strictly speaking unless all attributes (e.g. followers and friends) from the two lists are identical (which can only happen if both lists are taken at the same time), it is unlikely we can use "intersect" to find the users that are in both lists. As long as the "user" identifying attributes (e.g. screen_name) match, they should be considered the same user.